Move EuiRangeObject under components/form/range to match EUI's source layout
