+++ packages/test-helpers/src/index.ts
@@ -13,5 +13,5 @@ export { EuiSuperSelectObject } from './playwright/components/form/super_select/
 export { EuiGlobalToastListObject } from './playwright/components/toast/object';
 export { EuiSelectableObject } from './playwright/components/selectable/object';
 export { EuiFilterButtonObject } from './playwright/components/filter_button/object';
-export { EuiRangeObject } from './playwright/components/range/object';
+export { EuiRangeObject } from './playwright/components/form/range/object';
 export { EuiBasicTableObject } from './playwright/components/basic_table/object';
